Logout when we exit
This means we can re-login with the same hostname without changing
anything else.

I'm not certain context.Background() is correct, but it needs a context.
This should only be problematic if it takes a while to communicate to
Tailscale that we logged out, but Fly.io kills the service before that
completes.